added numerical scheme to approximate step width for membrane transport
currently not integrable with diffusion until the approach in unified
had to update epochs to long and reflect changing time to elapsed time

+++ singa-simulation/src/main/java/de/bioforscher/singa/simulation/events/EpochUpdateWriter.java
@@ -125,7 +125,7 @@ public class EpochUpdateWriter implements UpdateEventListener<NodeUpdatedEvent> 
     @Override
     public void onEventReceived(NodeUpdatedEvent event) {
         StringBuilder sb = new StringBuilder();
-        sb.append(Integer.toString(event.getEpoch())).append(SEPARATOR_CHARACTER);
+        sb.append(Long.toString(event.getEpoch())).append(SEPARATOR_CHARACTER);
         int count = 0;
         for (ChemicalEntity entity : this.observedEntities) {
             if (count < this.observedEntities.size() - 1) {
